chore(lint): cover more cases for noUndeclaredDependencies
