+++ crates/biome_js_analyze/src/analyzers/nursery/no_restricted_imports.rs
@@ -2,12 +2,10 @@ use biome_analyze::context::RuleContext;
 use biome_analyze::{declare_rule, Ast, Rule, RuleDiagnostic, RuleSource};
 use biome_console::markup;
 use biome_deserialize_macros::Deserializable;
-use biome_js_syntax::inner_string_text;
+use biome_js_syntax::{inner_string_text, AnyJsImportSpecifierLike};
 use biome_rowan::TextRange;
 use rustc_hash::FxHashMap;
 use serde::{Deserialize, Serialize};
-
-use super::use_nodejs_import_protocol::AnyJsImportLike;
 
 declare_rule! {
     /// Disallow specified modules when loaded by import or require.
@@ -45,7 +43,7 @@ pub struct RestrictedImportsOptions {
 }
 
 impl Rule for NoRestrictedImports {
-    type Query = Ast<AnyJsImportLike>;
+    type Query = Ast<AnyJsImportSpecifierLike>;
     type State = (TextRange, String);
     type Signals = Option<Self::State>;
     type Options = RestrictedImportsOptions;
