+++ crates/biome_js_analyze/tests/specs/nursery/noUndeclaredDependencies/invalid.js
@@ -1,1 +1,3 @@
 import "notInstalled";
+import("notInstalled");
+require("notInstalled")
