+++ crates/biome_js_analyze/tests/specs/nursery/noUndeclaredDependencies/valid.js
@@ -1,3 +1,9 @@
 import "./valid";
 import "react";
 import "@testing-library/react";
+import("./valid");
+import("react");
+import("@testing-library/react");
+require("./valid");
+require("react");
+require("@testing-library/react");
