+++ crates/biome_js_syntax/src/import_ext.rs
@@ -1,8 +1,9 @@
 use crate::{
-    inner_string_text, AnyJsBinding, AnyJsImportClause, AnyJsNamedImportSpecifier, JsImport,
-    JsImportAssertion, JsModuleSource, JsSyntaxToken,
+    inner_string_text, AnyJsBinding, AnyJsImportClause, AnyJsNamedImportSpecifier,
+    JsCallExpression, JsImport, JsImportAssertion, JsImportCallExpression, JsModuleSource,
+    JsSyntaxToken,
 };
-use biome_rowan::{AstNode, SyntaxResult, TokenText};
+use biome_rowan::{declare_node_union, AstNode, SyntaxResult, TokenText};
 
 impl JsImport {
     /// It checks if the source of an import against the string `source_to_check`
@@ -189,5 +190,118 @@ impl JsModuleSource {
     /// ```
     pub fn inner_string_text(&self) -> SyntaxResult<TokenText> {
         Ok(inner_string_text(&self.value_token()?))
+    }
+}
+
+declare_node_union! {
+    /// This node union is meant to match the following syntax:
+    /// ```js
+    ///    import "lodash";
+    /// //        ^^^^^^^^
+    ///    require("lodash")
+    /// // ^^^^^^^^^^^^^^^^^
+    ///    import("lodash")
+    /// // ^^^^^^^^^^^^^^^^
+    /// ```
+    pub AnyJsImportSpecifierLike = JsModuleSource | JsCallExpression |  JsImportCallExpression
+}
+
+impl AnyJsImportSpecifierLike {
+    /// Returns the inner text of specifier:
+    ///
+    /// ## Examples
+    ///
+    /// ```
+    /// use biome_js_factory::make;
+    /// use biome_js_syntax::AnyJsImportSpecifierLike;
+    /// use biome_rowan::TriviaPieceKind;
+    ///
+    /// let source_name = make::js_module_source(make::js_string_literal("foo"));
+    /// let any_import_specifier = AnyJsImportSpecifierLike::JsModuleSource(source_name);
+    /// assert_eq!(any_import_specifier.inner_string_text().unwrap().text(), "foo")
+    /// ```
+    pub fn inner_string_text(&self) -> Option<TokenText> {
+        match self {
+            AnyJsImportSpecifierLike::JsModuleSource(source) => source.inner_string_text().ok(),
+            AnyJsImportSpecifierLike::JsCallExpression(expression) => {
+                let callee = expression.callee().ok()?;
+                let name = callee.as_js_reference_identifier()?.value_token().ok()?;
+                if name.text_trimmed() == "require" {
+                    let [Some(argument)] = expression.arguments().ok()?.get_arguments_by_index([0])
+                    else {
+                        return None;
+                    };
+                    argument
+                        .as_any_js_expression()?
+                        .as_any_js_literal_expression()?
+                        .as_js_string_literal_expression()?
+                        .inner_string_text()
+                        .ok()
+                } else {
+                    None
+                }
+            }
+            AnyJsImportSpecifierLike::JsImportCallExpression(import_call) => {
+                let [Some(argument)] = import_call.arguments().ok()?.get_arguments_by_index([0])
+                else {
+                    return None;
+                };
+                argument
+                    .as_any_js_expression()?
+                    .as_any_js_literal_expression()?
+                    .as_js_string_literal_expression()?
+                    .inner_string_text()
+                    .ok()
+            }
+        }
+    }
+
+    /// Returns the whole token text of the specifier, with quotes included:
+    ///
+    /// ## Examples
+    ///
+    /// ```
+    /// use biome_js_factory::make;
+    /// use biome_js_syntax::AnyJsImportSpecifierLike;
+    /// use biome_rowan::TriviaPieceKind;
+    ///
+    /// let source_name = make::js_module_source(make::js_string_literal("foo"));
+    /// let any_import_specifier = AnyJsImportSpecifierLike::JsModuleSource(source_name);
+    /// assert_eq!(any_import_specifier.module_name_token().unwrap().text(), "\"foo\"")
+    /// ```
+    pub fn module_name_token(&self) -> Option<JsSyntaxToken> {
+        match self {
+            AnyJsImportSpecifierLike::JsModuleSource(source) => source.value_token().ok(),
+            AnyJsImportSpecifierLike::JsCallExpression(expression) => {
+                let callee = expression.callee().ok()?;
+                let name = callee.as_js_reference_identifier()?.value_token().ok()?;
+                if name.text_trimmed() == "require" {
+                    let [Some(argument)] = expression.arguments().ok()?.get_arguments_by_index([0])
+                    else {
+                        return None;
+                    };
+                    argument
+                        .as_any_js_expression()?
+                        .as_any_js_literal_expression()?
+                        .as_js_string_literal_expression()?
+                        .value_token()
+                        .ok()
+                } else {
+                    None
+                }
+            }
+            AnyJsImportSpecifierLike::JsImportCallExpression(import_call) => {
+                let [Some(argument)] = import_call.arguments().ok()?.get_arguments_by_index([0])
+                else {
+                    return None;
+                };
+                argument
+                    .as_any_js_expression()?
+                    .as_any_js_literal_expression()?
+                    .as_js_string_literal_expression()?
+                    .value_token()
+                    .ok()
+            }
+        }
     }
 }
