+++ crates/biome_js_syntax/src/lib.rs
@@ -34,6 +34,7 @@ pub use expr_ext::*;
 pub use file_source::*;
 pub use function_ext::*;
 pub use identifier_ext::*;
+pub use import_ext::*;
 pub use modifier_ext::*;
 pub use stmt_ext::*;
 pub use syntax_node::*;
